Added Comment Field
Added comment field to Lichen table.  Default image is populating, but not responsive to changes in selected model just yet.  To load images when a new licheen is selected, the URI must be generated in the delegate and assigned directly to the image element in either the ViewModel or the code-behind.